.history/requirements_20260322180302.txt
numpy
Pillow
streamlit
matplotlib
tensorflow-cpu==2.13.1
